+++ experiments/server_handling/pubspec.yaml
@@ -13,3 +13,4 @@ dev_dependencies:
   build_runner: ^2.1.4
   build_web_compilers: ^3.2.1
   lints: ^1.0.0
+  webdev: ^2.7.10
