+++ experiments/shelf_handler/lib/main.dart
@@ -1,20 +1,27 @@
 import 'dart:io';
 
-// server specific import
 import 'package:jaspr/jaspr_server.dart';
 import 'package:shelf/shelf_io.dart' as shelf_io;
+import 'package:shelf_router/shelf_router.dart';
 
-import 'app.dart';
+import 'components/app.dart';
+import 'components/hello.dart';
 
 void main() async {
   HttpServer? server;
 
-  var handler = const Pipeline() //
-      .addMiddleware(logRequests())
-      .addHandler(serveApp((request, render) {
+  var router = Router(notFoundHandler: serveApp((request, render) {
     print("Request uri is ${request.requestedUri}");
     return render(App());
   }));
+
+  router.get('/hello', (request) {
+    return renderComponent(Hello());
+  });
+
+  var handler = const Pipeline() //
+      .addMiddleware(logRequests())
+      .addHandler(router);
 
   server = await shelf_io.serve(handler, InternetAddress.anyIPv4, 8080);
