+++ experiments/shelf_handler/pubspec.yaml
@@ -9,6 +9,7 @@ environment:
 dependencies:
   jaspr: ^0.1.0
   shelf: ^1.3.1
+  shelf_router: ^1.1.3
 
 dev_dependencies:
   build_runner: ^2.1.4
