+++ experiments/shelf_handler/web/main.dart
@@ -1,5 +1,5 @@
 import 'package:jaspr/jaspr.dart';
-import 'package:server_handling_experiment/app.dart';
+import 'package:server_handling_experiment/components/app.dart';
 
 void main() {
   runApp(App());
