+++ packages/jaspr/lib/jaspr_server.dart
@@ -2,3 +2,4 @@ export 'package:shelf/shelf.dart';
 
 export 'jaspr.dart' hide runApp, AppBinding;
 export 'src/bindings/server_bindings.dart';
+export 'src/server/server_app.dart';
